Styling and UI auf 3. Seite: Styling und Seite Korrigieren: 1
Made final working structure for html so that the logic can work with the objects that are now on the page because of the design.
Also made structure so that "Anpassung der Seite bei Knopfdruck" Task could be implemented.
Only logic for creating output result has to be implemented correctly since right now it doesn't even use the anzahJahre object and only anzahlMonate object.